Add annotated photo of ADG621 chip
This chip is located on the bottom right of the back side of the board
(MSOP-10, marking SXB) and is a dual SPST switch. However, it does not
appear on Digilent's partial schematics and seems to be connected to
unpopulated resistor footprints that would connect it to other
components. I haven't mapped out enough connections to tell what this
was supposed to be used for.