Add boundary scan and analysis for FT232H GPIOs
Looking at the board, it appears that there are traces going from the
FT232H GPIO pins to the FPGA. By configuring these as outputs, setting
them high one by one, and dumping the boundary scan register the mapping
can be found. GPIOH7 seems to be used to control the output enables on
the NC7SZ126 chips.