Add boundary scan and analysis ADG621 IN1/2 pin
I was previously probing the side of the unpopulated resistor that
wasn't connected to the IN1/2 pins of the ADG621, but wasn't noticing
any changes in the boundary scan. By probing the side connected to the
IC, a bit did change in the boundary scan. The unpopulated footprint
seems to be for a pull-up resistor to 3.3V.