wires: Basic VHDL design to connect unmapped pins to AD2 DIO lines
This is a really basic design, essentially just wires between the 8
currently unmapped FPGA pins and DIO pins 0-7 on the Analog Discovery,
which were previously mapped. It also wires the LED to DIO8 for
convenience.

Since the FPGA cannot use its JTAG pins for (high speed) IO, it must use
regular IO lines when communicating to the FT232H. However, since the
first 4 IO lines on the FT232H are used for both JTAG and FIFO modes,
they must be shared between the IO and JTAG lines of the FPGA. To do
this, there are some tri-state buffer chips on TMS and TDO to prevent
the two data streams from interfering with each other. Since TCK and TDI
are inputs to the FPGA, and the JTAG state will not change because of
them, they are not buffered. Because of this architecture, it is not
possible to run boundary scans to map out these pins. However, by
connecting these unmapped pins to externally accessible pins, those 4
FT232H pins can be mapped out by setting them high one by one in bitbang
mode, and seeing which external pin goes high as a result. An easy way
of checking this is by feeding the external pins being tested to the
onboard LED with a jumper wire.