Add FPGA mapping for more FT232H pins
Using a basic FPGA design, the ADBUS0-3 on the FT232H can be mapped out.
The ftdi-bitbang [1] program was used to set each ADBUS pin high one by
one. It also seems like ACBUS7 (GPIOH7) on the FT232H is also connected
to the FPGA. This pin is used to control the output enable signals on
the tri-state buffers, and is likely used by Digilent's bitstreams to
indicate when to stop using the shared IO lines due to the FT232H
switching to JTAG mode. Since the first 4 ADBUS pins aren't used as
GPIOs in MPSSE mode, rename the FT232H pins in the mapping document to
the more generic pin names.

[1]: https://github.com/aehparta/ftdi-bitbang